+++ notification-service/src/main/java/com/flight_booking/notification_service/application/service/NotificationService.java
@@ -1,10 +1,10 @@
-package com.flight_booking.notification_service.service;
+package com.flight_booking.notification_service.application.service;
 
-import com.flight_booking.notification_service.dto.NotificationRequest;
-import com.flight_booking.notification_service.dto.NotificationResponse;
-import com.flight_booking.notification_service.exception.NotificationNotFoundException;
-import com.flight_booking.notification_service.model.Notification;
-import com.flight_booking.notification_service.repository.NotificationRepository;
+import com.flight_booking.notification_service.presentation.dto.NotificationRequest;
+import com.flight_booking.notification_service.presentation.dto.NotificationResponse;
+import com.flight_booking.notification_service.global.exception.NotificationNotFoundException;
+import com.flight_booking.notification_service.domain.model.Notification;
+import com.flight_booking.notification_service.domain.repository.NotificationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
