+++ notification-service/src/main/java/com/flight_booking/notification_service/domain/model/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.flight_booking.notification_service.model;
+package com.flight_booking.notification_service.domain.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
