+++ notification-service/src/main/java/com/flight_booking/notification_service/global/exception/NotificationNotFoundException.java
@@ -1,4 +1,4 @@
-package com.flight_booking.notification_service.exception;
+package com.flight_booking.notification_service.global.exception;
 
 public class NotificationNotFoundException extends RuntimeException {
   public NotificationNotFoundException(String message) {
