+++ notification-service/src/main/java/com/flight_booking/notification_service/presentation/dto/NotificationResponse.java
@@ -1,4 +1,4 @@
-package com.flight_booking.notification_service.dto;
+package com.flight_booking.notification_service.presentation.dto;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
