+++ notification-service/src/test/java/com/flight_booking/notification_service/EmailServiceImplTest.java
@@ -1,8 +1,8 @@
 package com.flight_booking.notification_service;
 
-import com.flight_booking.notification_service.model.Notification;
-import com.flight_booking.notification_service.repository.NotificationRepository;
-import com.flight_booking.notification_service.service.EmailServiceImpl;
+import com.flight_booking.notification_service.domain.model.Notification;
+import com.flight_booking.notification_service.domain.repository.NotificationRepository;
+import com.flight_booking.notification_service.application.service.EmailServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
